Squashed commit of the following:
commit 43a7cc559df44d46999070b7318578b29b568900
Author: bri25yu <bri25yu@gmail.com>
Date:   Sat Oct 3 16:01:03 2020 -0700

    moving everything into one column

commit d7f084b047914c5d1746b3779ee6487765663c1d
Author: bri25yu <bri25yu@gmail.com>
Date:   Sat Oct 3 15:09:04 2020 -0700

    refactoring dataflow of candidate portal checklist

commit 4ff9ef1aefba37fd6422a7a41cf7d5b71e530189
Author: bri25yu <bri25yu@gmail.com>
Date:   Fri Oct 2 23:37:33 2020 -0700

    refactoring candidate checklist events

commit fa0d04e2b9610882f679e1a68abeb79bae848adc
Author: bri25yu <bri25yu@gmail.com>
Date:   Fri Oct 2 23:27:47 2020 -0700

    refactoring candidate portal checklist titles

commit 8a3fa43d7c0399bd836ddc128b33764a684fdd46
Author: bri25yu <bri25yu@gmail.com>
Date:   Fri Oct 2 23:01:51 2020 -0700

    refactoring candidate classes into utils

commit 6f366a04b94416415eb1220dbbaca466e473d9a8
Author: bri25yu <bri25yu@gmail.com>
Date:   Fri Oct 2 22:44:39 2020 -0700

    organizing candidate imports

commit ac7c31cd10483847d36db731051e3af0cfd54b45
Author: bri25yu <bri25yu@gmail.com>
Date:   Fri Oct 2 22:40:07 2020 -0700

    removing candidate/tests since it's empty
